Ignore resume on cancelled continuation
There is a race between cancellation of outer job and invocation of
resume on a cancellable continuation and this race cannot be
worked around by the code that uses suspendCancellableCoroutine
function. Thus resume attempt on a cancelled continuation shall
be simply ignored.

Fixes #450
